Main Activity changed to HomeFragment
