+++ app/src/main/AndroidManifest.xml
@@ -20,25 +20,28 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.BocchiTv">
         <activity
-            android:name=".DetailsPage.DetailsActivity"
-            android:exported="false">
-        </activity>
-        <activity
             android:name=".MainPage.MainActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LEANBACK_LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
-            android:name=".NetworkActivity"
-            android:exported="false">
+            android:name=".DetailsPage.DetailsActivity"
+            android:exported="false" />
+        <activity
+            android:name=".MainPage.HomeFragment"
+            android:exported="true">
+
         </activity>
         <activity
+            android:name=".NetworkActivity"
+            android:exported="false" />
+        <activity
             android:name=".VideoPlayerActivity"
-            android:exported="false">
-        </activity>
+            android:exported="false" />
     </application>
 
 </manifest>
