+++ app/src/main/res/layout/activity_main.xml
@@ -1,100 +1,61 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainPage.MainActivity">
 
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/top_guideline"
+    <FrameLayout
+        android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:orientation="horizontal"
-        app:layout_constraintGuide_percent="0.1" />
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/side_guideline"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        app:layout_constraintGuide_percent="0.06" />
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/center_guideline"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal"
-        app:layout_constraintGuide_percent="0.5" />
-
-    <ImageView
-        android:id="@+id/img_banner"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:adjustViewBounds="true"
-        app:layout_constraintDimensionRatio="16:9"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".6" />
-
-    <View
-        android:id="@+id/gradient_horizontal"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:background="@drawable/banner_gradient"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="@id/side_guideline"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".7" />
-
-    <TextView
-        android:layout_width="0dp"
-        android:id="@+id/title"
-        android:layout_height="wrap_content"
-        android:textColor="#fff"
-        android:textSize="26sp"
-        app:layout_constraintStart_toStartOf="@+id/side_guideline"
-        app:layout_constraintTop_toTopOf="@+id/top_guideline"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".4"
-        android:text="Title" />
-    <TextView
-        android:id="@+id/subtitle"
-        android:layout_width="wrap_content"
-        android:text="Language : English"
-        android:textSize="14sp"
-        android:textColor="#fff"
-        android:layout_marginTop="16dp"
-        app:layout_constraintTop_toBottomOf="@+id/title"
-        app:layout_constraintStart_toStartOf="@+id/side_guideline"
-        android:layout_height="wrap_content"/>
-
-    <TextView
-        android:id="@+id/description"
-        android:layout_width="0dp"
-        android:text="Language : English"
-        android:textSize="13sp"
-        android:textColor="#fff"
-        android:layout_marginTop="16dp"
-        android:maxLines="4"
-        android:lineSpacingMultiplier="1.2"
-        app:layout_constraintWidth_default="percent"
-        app:layout_constraintWidth_percent=".4"
-        app:layout_constraintTop_toBottomOf="@+id/subtitle"
-        app:layout_constraintStart_toStartOf="@+id/side_guideline"
-        android:layout_height="wrap_content"/>
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/list_fragment"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/center_guideline"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:nextFocusRight="@id/blfNavBar"
         />
+    <androidx.leanback.widget.BrowseFrameLayout
+        android:id="@+id/blfNavBar"
+        android:layout_width="50dp"
+        android:layout_height="match_parent"
+        android:background="#000000"
+        android:nextFocusLeft="@id/container"
+        >
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_marginTop="32dp"
+            android:orientation="vertical">
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="45dp"
+                android:id="@+id/btn_search"
+                android:focusable="true"
+                />
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="45dp"
+                android:id="@+id/btn_home"
+                android:focusable="true"
+                />
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="45dp"
+                android:id="@+id/btn_movies"
+                android:focusable="true"
+                />
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="45dp"
+                android:id="@+id/btn_tv"
+                android:focusable="true"
+                />     <TextView
+            android:layout_width="match_parent"
+            android:layout_height="45dp"
+            android:id="@+id/btn_sports"
+            android:focusable="true"
+            />
 
+        </LinearLayout>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+    </androidx.leanback.widget.BrowseFrameLayout>
+
+</merge>
